+++ packages/components/atoms/f-popover/test/specs/accessibility/axe-accessibility.spec.js
@@ -1,0 +1,8 @@
+import { getAccessibilityTestResults } from '../../../../../../../test/utils/axe-helper';
+
+describe('Accessibility tests', () => {
+    it('a11y - should test f-popover component WCAG compliance', () => {
+        // Act
+        const axeResults = getAccessibilityTestResults('f-popover');
+    });
+});