+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -175,32 +175,6 @@ export default {
             default: 1000
         },
 
-        getCheckoutTimeout: {
-            type: Number,
-            required: false,
-            default: 1000
-        },
-
-        createGuestTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        getBasketTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        placeOrderTimeout: {
-            type: Number,
-            default: 1000
-        },
-
-        updateCheckoutTimeout: {
-            type: Number,
-            default: 1000
-        },
-
         authToken: {
             type: String,
             default: ''
@@ -216,12 +190,12 @@ export default {
             required: true
         },
 
-        getAddressTimeout: {
-            type: Number,
-            default: 1000
+        applicationName: {
+            type: String,
+            required: true
         },
 
-        applicationName: {
+        getGeoLocationUrl: {
             type: String,
             required: true
         }
@@ -271,7 +245,8 @@ export default {
             'time',
             'userNote',
             'basket',
-            'orderId'
+            'orderId',
+            'geolocation'
         ]),
 
         isMobileNumberValid () {
@@ -329,7 +304,8 @@ export default {
             'setAuthToken',
             'updateCustomerDetails',
             'updateUserNote',
-            'placeOrder'
+            'placeOrder',
+            'getGeoLocation'
         ]),
 
         ...mapActions('analytics', [
@@ -374,18 +350,21 @@ export default {
                     await this.setupGuestUser();
                 }
 
+                await this.lookupGeoLocation();
+
                 const data = mapUpdateCheckoutRequest({
                     address: this.address,
                     customer: this.customer,
                     isCheckoutMethodDelivery: this.isCheckoutMethodDelivery,
                     time: this.time,
-                    userNote: this.userNote
+                    userNote: this.userNote,
+                    geolocation: this.geolocation
                 });
 
                 await this.updateCheckout({
                     url: this.updateCheckoutUrl,
                     data,
-                    timeout: this.updateCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 await this.submitOrder();
@@ -441,7 +420,7 @@ export default {
             await this.placeOrder({
                 url: this.placeOrderUrl,
                 data,
-                timeout: this.placeOrderTimeout
+                timeout: this.checkoutTimeout
             });
         },
 
@@ -462,7 +441,7 @@ export default {
                     url: this.createGuestUrl,
                     tenant: this.tenant,
                     data: createGuestData,
-                    timeout: this.createGuestTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutSetupGuestSuccess);
@@ -485,7 +464,7 @@ export default {
             try {
                 await this.getCheckout({
                     url: this.getCheckoutUrl,
-                    timeout: this.getCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutGetSuccess);
@@ -511,7 +490,7 @@ export default {
                     url: this.getBasketUrl,
                     tenant: this.tenant,
                     language: this.$i18n.locale,
-                    timeout: this.getBasketTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutBasketGetSuccess);
@@ -535,7 +514,7 @@ export default {
             try {
                 await this.getAvailableFulfilment({
                     url: this.checkoutAvailableFulfilmentUrl,
-                    timeout: this.getCheckoutTimeout
+                    timeout: this.checkoutTimeout
                 });
 
                 this.$emit(EventNames.CheckoutAvailableFulfilmentGetSuccess);
@@ -561,13 +540,39 @@ export default {
                     url: this.getAddressUrl,
                     tenant: this.tenant,
                     language: this.$i18n.locale,
-                    timeout: this.getAddressTimeout
+                    timeout: this.checkoutTimeout
                 });
-
                 this.$emit(EventNames.CheckoutAddressGetSuccess);
             } catch (thrownErrors) {
                 this.$emit(EventNames.CheckoutAddressGetFailure, thrownErrors);
                 this.hasCheckoutLoadedSuccessfully = false;
+            }
+        },
+
+        /**
+         * Look up the geo details for the customers address, skip on failure.
+         *
+         */
+        async lookupGeoLocation () {
+            const locationData =
+            {
+                addressLines: Object.values(this.address).filter(addressLine => !!addressLine)
+            };
+
+            try {
+                if (locationData.addressLines.length) {
+                    await this.getGeoLocation({
+                        url: this.getGeoLocationUrl,
+                        postData: locationData,
+                        timeout: this.checkoutTimeout
+                    });
+                }
+            } catch (error) {
+                this.$logger.logWarn(
+                    'Geo location lookup failed',
+                    this.$store,
+                    { error }
+                );
             }
         },
 
@@ -745,7 +750,7 @@ export default {
     .c-spinner {
         margin: 0 auto;
     }
-}
+  }
 
 .c-checkout {
     padding-top: spacing(x6);